requirements.txt
Flask
Flask-CORS
tensorflow
Pillow
numpy
python-dotenv
